// front/api-dentysalud-ts/src/citas/cita.model.d.ts
export interface CitaDTO{
    idcita: interger;
    nombrepaciente:string;
    apellidopaciente:string;
    local:strging;
    hora:integer;
    fecha:string;
}

export interface CitaRegistrarDTO{
    nombrepaciente:string;
    apellidopaciente:string;
    local:strging;
    hora:integer;
    fecha:string;
}